+++ src/templates/home/index.js
@@ -1,0 +1,13 @@
+import { HomeContainer } from "./home";
+import { ProjectsContainer } from "./projects";
+import { LestTalkContainer } from "./lestTalk";
+
+export function HomeTemplate() {
+  return (
+    <>
+      <HomeContainer />
+      <ProjectsContainer />
+      <LestTalkContainer />
+    </>
+  );
+}
